Remove typescript, tslint, change all file extensions
