+++ package.json
@@ -37,9 +37,9 @@
     "react-router": "^4.2.0",
     "react-router-dom": "^4.2.2",
     "react-scripts": "2.1.3",
-    "react-scripts-ts": "2.16.0",
     "recharts": "^1.0.0-beta.10",
     "redux": "^3.7.2",
+    "redux-logger": "^3.0.6",
     "redux-saga": "^0.16.2",
     "redux-thunk": "2.2.0",
     "reselect": "^3.0.1",
@@ -64,10 +64,7 @@
     "eslint-plugin-prettier": "^3.0.1",
     "eslint-plugin-react": "^7.12.3",
     "prettier": "^1.15.3",
-    "react-test-renderer": "^16.7.0",
-    "redux-logger": "^3.0.6",
-    "tslint": "^5.10.0",
-    "typescript": "^2.9.2"
+    "react-test-renderer": "^16.7.0"
   },
   "browserslist": [
     ">0.2%",
