+++ src/index.jsx
@@ -1,10 +1,9 @@
-import * as React from 'react';
-
+import React from 'react';
 import { render } from 'react-dom';
-import 'bootstrap/dist/css/bootstrap.min.css';
-import configureStore from './store/configureStore';
 import { Provider } from 'react-redux';
+import configureStore from './store/configureStore';
 import App from './components/App/App';
+import 'bootstrap/dist/css/bootstrap.min.css';
 import './index.css';
 
 const store = configureStore();
