+++ src/reducers/currentDevice.js
