+++ src/reducers/deviceDetail.js
