+++ src/reducers/deviceList.js
