+++ src/reducers/devices.js
