+++ src/reducers/properties.js
