+++ src/selectors/commandOutput.js
