+++ src/selectors/currentDevice.js
