+++ src/selectors/deviceDetail.js
