+++ src/selectors/deviceList.js
